% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_palette.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Gets the color palette of an image as a tibble}
\usage{
get_color_palette(img_url, limit)
}
\arguments{
\item{img_url}{a character string representing the image URL}

\item{limit}{the maximum number of colors to be returned}
}
\value{
data.frame (tibble::tibble)
}
\description{
Extracts the most common colors, configurable through parameters, from
an image specified as a URL and returns them as a data frame of hex
color codes and RGB values. Includes the share of each color in the image.
}
\examples{
url <- "https://visit.ubc.ca/wp-content/uploads/2019/04/plantrip_header-2800x1000_2x.jpg"
get_color_palette(url, 3)
}
