% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative.R
\name{negative}
\alias{negative}
\title{Invert top n colours in an image file.}
\usage{
negative(img_url, num_colours, threshold)
}
\arguments{
\item{img_url}{URL of an image file}

\item{num_colours}{number of colours to be extracted}

\item{threshold}{minimum share of all pixels; between 0 and 1}
}
\value{
data.frame (tibble::tibble)
}
\description{
(e.g. red becomes cyan, green becomes magenta, yellow becomes blue)
then stored in a table as HEX codes and RGB values.
}
\examples{
url <- "https://visit.ubc.ca/wp-content/uploads/2019/04/plantrip_header-2800x1000_2x.jpg"
negative(url, 3, 0.2)
}
