% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative.R
\name{rgb2col}
\alias{rgb2col}
\title{Converts a column of RGB values to a HEX code
Based on: https://stackoverflow.com/a/66810164}
\usage{
rgb2col(mat, col)
}
\arguments{
\item{mat}{matrix of RGB color codes}

\item{col}{column containing RGB values; row order is red, green, blue}
}
\value{
character vector of the HEX code
}
\description{
Converts a column of RGB values to a HEX code
Based on: https://stackoverflow.com/a/66810164
}
\examples{
rgb2col(grDevices::col2rgb(c("#000C1C")))
}
