% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Create a two-dimensional scatterplot based on the colours of the image}
\usage{
scatterplot(url, dataset, x, y, colour, tolerance)
}
\arguments{
\item{dataset}{data.frame or tibble}

\item{x}{unquoted column name to plot on the x-axis from data data.frame or tibble}

\item{y}{unquoted column name to plot on the y-axis from data data.frame or tibble}

\item{colour}{unquoted column name to plot as the colour from data data.frame or tibble}

\item{tolerance}{threshold minimum share of all pixels; between 0 and 1}

\item{img_url}{URL of an image file}
}
\value{
A geom_point ggplot
}
\description{
Creates a simple scatterplot using the colours selected from the image,
plotting two features from a dataset of the users choosing.
}
\examples{
scatterplot('https://visit.ubc.ca/wp-content/uploads/2019/04/plantrip_header-2800x1000_2x.jpg', penguins, bill_length_mm, body_mass_g, species, 50)
}
