% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_palette.R
\name{url_exists}
\alias{url_exists}
\title{Checks if a URL exists}
\usage{
url_exists(url)
}
\arguments{
\item{url}{URL}
}
\value{
True if exists
}
\description{
Check if the specified URL returns a valid response (status code 200)
}
\examples{
url_exists("https://www.google.com")
}
