% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative.R
\name{negative}
\alias{negative}
\title{Invert top n colours in an image file.}
\usage{
negative(img_url, num_colours)
}
\arguments{
\item{img_url}{URL of an image file}

\item{num_colours}{number of colours to be extracted}
}
\value{
data.frame (tibble::tibble)
}
\description{
Colours are extracted from an image via URL and reversed,
(e.g. red becomes cyan, green becomes magenta, yellow becomes blue)
then stored in a table as HEX codes and RGB values.
}
\examples{
url <- "https://i.imgur.com/s9egWBB.jpg"
negative(url, 3)
}
