% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative.R
\name{negative}
\alias{negative}
\title{Invert top n colours in an image file.}
\usage{
negative(img_url, num_colours, threshold = 0)
}
\arguments{
\item{img_url}{URL of an image file}

\item{num_colours}{number of colours to be extracted}

\item{threshold}{minimum share of all pixels; between 0 and 1; 0 by default}
}
\value{
data.frame (tibble::tibble)
}
\description{
Colours are extracted from an image via URL and reversed,
(e.g. red becomes cyan, green becomes magenta, yellow becomes blue)
then stored in a table as HEX codes and RGB values.
}
\examples{
url <- "https://masterdatascience.ubc.ca/sites/default/files/styles/banner_image_large/public/media-images/fos-datascience-healthineers_1440x512-fina-webl.jpg?itok=rBc6RLF6"
negative(url, 3, 0.001)
}
