% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Create a two-dimensional scatterplot based on the colours of the image}
\usage{
scatterplot(img_url, dataset, x, y, colour)
}
\arguments{
\item{img_url}{URL of an image file}

\item{dataset}{data.frame or tibble}

\item{x}{unquoted column name to plot on the x-axis from data data.frame or tibble}

\item{y}{unquoted column name to plot on the y-axis from data data.frame or tibble}

\item{colour}{unquoted column name to plot as the colour from data data.frame or tibble}
}
\value{
A geom_point ggplot
}
\description{
Creates a simple scatterplot using the colours selected from the image,
plotting two features from a dataset of the users choosing.
}
\examples{
scatterplot('https://i.imgur.com/s9egWBB.jpg',
palmerpenguins::penguins,
bill_length_mm, body_mass_g, species)
}
